feat: remove unnecessary descriptor field
There's no real reason for src to be in the descriptor. It is there
only as a convenience to make the webpack configuration easier, but
that's an implementation detail that doesn't need to be exposed.  It
is also only needed for widget module federation, so remove it from
the generic ItemDescriptor.