fix: don't target node
Setting target to "node" causes webpack to include some require()
statements in remoteEntry.js that fail in a browser environment.
For payment gateways, it's sufficient to set the output library
type to commonjs2 so they can be imported in a node environment.
Widgets will continue to be exported in remoteEntry.js using
library type "window" as configured by ModuleFederationPlugin.