fix: don't capture user in authentication context
With react-oidc-context, whenever the access token is refreshed,
it causes both the auth state (oidcInstance) and the user to
change.  If this is captured in the authentication context, then
it means the authentication context is changed as well, which
then cascades to api services since the latter captures the
authentication context in order to set the Authorization header.
Depending on usage, an api services update may inadvertently
trigger an unnecessary useEffect (yes the access token changed,
but that's no reason to trigger a new api call).

This commit uses a ref to get the current authenticated user in
lieu of capturing it in the authentication context.  Note that
with this implementation, if the user's name or email is updated
within a login session, it will not be updated in the authentication
context (for the same reason that we don't want to inadvertently
trigger unnecessary api calls).